coding_assessment_agent/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
psycopg2-binary
redis
langchain
langchain-openai
langchain-community
chromadb
openai
python-dotenv
pydantic
alembic
pytest
pytest-asyncio
httpx
greenlet
